scripts/在线人数/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alibabacloud-darabonba-string==0.0.4
alibabacloud-kms20160120==2.2.3
alibabacloud-tea==0.4.0
alibabacloud_credentials==0.3.6
alibabacloud_darabonba_array==0.1.0
alibabacloud_darabonba_encode_util==0.0.2
alibabacloud_darabonba_map==0.0.1
alibabacloud_darabonba_signature_util==0.0.4
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_pop==0.0.6
alibabacloud_gateway_spi==0.0.2
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.12
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Cython==3.0.12
Flask==3.1.0
frozenlist==1.5.0
grpcio==1.70.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
nacos-sdk-python==2.0.1
packaging==24.2
prometheus_client==0.21.1
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyOpenSSL==25.0.0
PyYAML==6.0.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
